fix : 형 변환 float -> double
Former-commit-id: 8bb0b87580865105ad4e5d0373f01e53ef25bcc3

+++ backend/api-server/src/main/java/com/dawool/api/dto/detailInfo/CommonInfoDto.java
@@ -37,9 +37,9 @@ public class CommonInfoDto {
     // 도로명 주소
     private String addr1;
     // 경도
-    private float mapX;
+    private double mapX;
     // 위도
-    private float mapY;
+    private double mapY;
     // 지도 레벨
     private float mLevel;
     // 청각장애
